Enable auto manifest updates
flatpak-external-data-checker (f-e-d-c) is run across Flathub hourly. When a new
tag is found in treesheets upstream, flathubbot will open a PR in this repo,
update the tag and commit in the manifest, and start test builds. You
still control when to merge and update the package on Flathub.

See https://docs.flathub.org/docs/for-app-authors/external-data-checker/